WIP: Optional feature: More verbose failed expression reporting
An option is added to the Parser class. This option is disabled by default,
and the existing behavior is fully preserved.

When the option is enabled, the final expected message is extended with
extra information about the previously "weakly failed" rules. This way,
not only the last failed NoMatch exception and its failing rules are displayed
but also all the rules that were not matched during the whole parsing process.