Fix PEG parsers' "comment" rule
The problem with the previous version was that if there is no text after
the '//', regular whitespace processing kicked in and the regexp was
applied to the next non-empty line.

The effect of this was that the next line gets commented off
unconditionally, which is kindof not what you want.